backend/requirements.txt
# AI Interviewer Backend Dependencies
# Python 3.10+ required

# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9

# AI & Speech
groq==0.9.0

# MongoDB Database
motor==3.3.2
pymongo==4.6.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
pydantic[email]==2.5.3

# Rate Limiting
slowapi==0.1.9

# Environment
python-dotenv==1.0.1

# Production Server (optional)
gunicorn==22.0.0
